debug: Remove astype string

+++ jobs/ml_jobs/algo_training/two_towers_model/utils/layers.py
@@ -48,8 +48,8 @@ class NumericalFeatureProcessor:
 
     def build_sequential_layer(self, vocabulary: np.ndarray):
         # Precompute mean/variance offline
-        mean = np.mean(vocabulary)
-        variance = np.var(vocabulary)
+        mean = np.mean(vocabulary.astype("float32"))
+        variance = np.var(vocabulary.astype("float32"))
 
         # Create and adapt the normalization layer to the data
         normalization = tf.keras.layers.Normalization(
